Cleaned up a lot of specs.

+++ spec/info_sqlite3.rb
@@ -1,20 +1,25 @@
 class Baza::InfoSqlite3
-  def self.sample_db
+  attr_reader :db
+
+  def initialize
     require "sqlite3"
-    path = "#{Dir.tmpdir}/baza_sqlite3_test.sqlite3_#{Time.now.to_f.to_s.hash}"
-    File.unlink(path) if File.exists?(path)
-    db = Baza::Db.new(
-      :type => :sqlite3,
-      :path => path,
-      :index_append_table_name => true,
-      :sql_to_error => true
+    require "tmpdir"
+
+    @path = "#{Dir.tmpdir}/baza_sqlite3_test_#{Time.now.to_f.to_s.hash}.sqlite3"
+    File.unlink(path) if File.exists?(@path)
+    @db = Baza::Db.new(
+      type: :sqlite3,
+      path: @path,
+      index_append_table_name: true,
+      sql_to_error: true
     )
-    
-    begin
-      yield db
-    ensure
-      db.close
-      File.unlink(path)
-    end
+  end
+
+  def before
+  end
+
+  def after
+    @db.close
+    File.unlink(@path)
   end
 end
